update auth service to validate mongoose id and password  and update package.json
validate mongoose id in admin-ban-users and admin-delete-ban and admin-delete-user route and validate required password in reset-password route

+++ auth/src/routes/admin-delete-users.ts
@@ -1,6 +1,7 @@
 import express, { Request, Response } from "express";
 import { User } from "../models/user.model";
 import { requireAuth, BadRequestError } from "@mesocial/common";
+import mongoose from "mongoose";
 
 const router = express.Router();
 
@@ -12,6 +13,10 @@ router.delete("/api/auth/admin",
 
         if (!user?.isAdmin) {
             throw new BadRequestError("User have no this permission");
+        }
+
+        if (!req.query.id || !mongoose.Types.ObjectId.isValid(String(req.query.id))) {
+            throw new BadRequestError("Id Is Invalid");
         }
 
         await User.findByIdAndDelete(req.query.id);
